+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Show GitHub ping webhook event on status dashboard instead of "No events received yet"
+- Return HTTP 202 instead of 400 for webhooks targeting a non-configured branch
+
 ## [[1.1.2](https://github.com/open-resource-discovery/provider-server/releases/tag/v1.1.2)] - 2026-04-23
 
 ### Added